% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findKeithRoot.R
\name{findKeithRoot}
\alias{findKeithRoot}
\title{Finds the root dir of a project made with keithUtils}
\usage{
findKeithRoot()
}
\value{
the root dir, defined as a folder containing
    ".projectInfo.RData". Throws an error if no root
    is found.
}
\description{
Finds the root dir of a project made with keithUtils
}
\examples{
findKeithRoot()
}
