% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.colors.R
\name{jet.colors}
\alias{jet.colors}
\title{Emulates the jet colormap from Matlab}
\usage{
jet.colors(n = 64)
}
\arguments{
\item{n}{the number of levels in the colormap;
the default is 64}
}
\value{
an rgb vector of length n
}
\description{
Emulates the jet colormap from Matlab
}
\examples{
plot(1:64, col=jet.colors(64))
}
