% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keithUtils.R
\docType{package}
\name{keithUtils}
\alias{keithUtils}
\alias{keithUtils-package}
\title{Utility functions for project assembly}
\description{
This package contains tools I've found useful
as I try to produce more reproducible reports
and workflows, as well as examples of specific
tricks I keep rediscovering (such as graphics
examples).
}
