% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMdaccReport.R
\name{newMdaccReport}
\alias{newMdaccReport}
\title{Build a new MDACC Report from a Template}
\usage{
newMdaccReport(reportFilename, reportPrefix = "default")
}
\arguments{
\item{reportFilename}{Ideally informative text name for
the report. This should not include a numerical
prefix, as that should be autogenerated. There is
no default here; you have to think of something.}

\item{reportPrefix}{If this is left as "default", the
current directory will be checked for all files
beginning with "r**_", where the "**" corresponds
to two numbers. The largest number currently present
will be incremented by one, and that new value will
be used. Can be overridden.}
}
\value{
invisible if the report subfolder and skeleton
    are successfully built.
}
\description{
This pulls in a template report skeleton from
the package and names it using a prefix numbering
which should reflect the order in which the reports
are written. New reports are placed in their own
subfolders. This is largely a wrapper for
rmarkdown::draft() with some prespecified arguments.
}
\examples{
newMdaccReport("preprocessData")
}
