% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProjectDescription.R
\name{newProjectDescription}
\alias{newProjectDescription}
\title{Build a new Project Description from a Template}
\usage{
newProjectDescription()
}
\value{
invisible if the report subfolder and skeleton
    are successfully built.
}
\description{
This pulls in a template description skeleton from
the package. There should just be one project description,
and it should reside at the top level of the project.
It should describe why we're undertaking this in broad
terms.
This is largely a wrapper for
rmarkdown::draft() with some prespecified arguments.
}
\examples{
newProjectDescription()
}
