% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setUpProject}
\alias{setUpProject}
\title{Autopopulates project folder with common directories}
\usage{
setUpProject(folderStyle = "KeithBaggerly",
  includeProjectDescription = TRUE)
}
\arguments{
\item{folderStyle}{character string naming the person
whose file structure I'm using as a template.
Default is "KeithBaggerly"; other currently allowed
variants are "JennyBryan", and "KarlBroman".}

\item{includeProjectDescription}{logical - should a
skeleton project description file be included at
the top level of the project hierarchy? Defaults
to TRUE.}
}
\value{
invisible if all of the folders were successfully
    created.
}
\description{
I almost always start a project by creating subdirectories
for RawData, ProcessedData, and so on, so I decided to
script the process to improve consistency.
}
\examples{
setupProject("KeithBaggerly")
}
